Split doc-related code into multiple files
