+++ docs/generator.py
@@ -1,38 +1,15 @@
-import argparse
 import importlib
-import json
 import os
-import re
-import shutil
 import sys
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
-from urllib import error as urllib_error
-from http import HTTPStatus
 
+import re
 import yaml
-from ansible.module_utils._text import to_text
-from ansible.module_utils.urls import open_url
 from jinja2 import Environment, FileSystemLoader
 
-from httpapi_plugins.ftd import BASE_HEADERS
 from module_utils.common import HTTPMethod, IDENTITY_PROPERTIES
-from module_utils.fdm_swagger_client import FdmSwaggerParser, SpecProp, OperationField, PropName, OperationParams, \
-    FILE_MODEL_NAME, QueryParams
-from module_utils.configuration import OperationNamePrefix, OperationChecker
-
-BASE_DIR_PATH = os.path.dirname(os.path.realpath(__file__))
-DEFAULT_TEMPLATE_DIR = os.path.join(BASE_DIR_PATH, 'templates')
-DEFAULT_SAMPLES_DIR = os.path.join(os.path.dirname(BASE_DIR_PATH), 'samples')
-DEFAULT_STATIC_DIR = os.path.join(BASE_DIR_PATH, 'static')
-DEFAULT_DIST_DIR = os.path.join(BASE_DIR_PATH, 'dist')
-DEFAULT_MODULE_DIR = os.path.join(os.path.dirname(BASE_DIR_PATH), 'library')
-
-SUPPORTED_VERSIONS = ['v2', 'v1']
-TOKEN_PATH_TEMPLATE = '/api/fdm/{}/fdm/token'
-
-SPEC_PATH = '/apispec/ngfw.json'
-DOC_PATH = '/apispec/en-us/doc.json'
+from module_utils.fdm_swagger_client import SpecProp, OperationField, PropName, OperationParams, FILE_MODEL_NAME
 
 ModelSpec = namedtuple('ModelSpec', 'name description properties operations')
 OperationSpec = namedtuple('OperationSpec', 'name description model_name path_params query_params data_params')
@@ -271,140 +248,3 @@ class ExampleDocGenerator(BaseDocGenerator):
 def camel_to_snake(text):
     test_with_underscores = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', text)
     return re.sub('([a-z0-9])([A-Z])', r'\1_\2', test_with_underscores).lower()
-
-
-def fetch_api_specs_with_docs(hostname, username, password):
-    def request_token(token_url):
-        resp = open_url(
-            hostname + token_url,
-            method=HTTPMethod.POST,
-            data=json.dumps({'grant_type': 'password', 'username': username, 'password': password}),
-            headers=BASE_HEADERS,
-            validate_certs=False
-        ).read()
-        return json.loads(to_text(resp))
-
-    def get_token():
-        for version in SUPPORTED_VERSIONS:
-            token_url = TOKEN_PATH_TEMPLATE.format(version)
-            try:
-                token = request_token(token_url)
-            except urllib_error.HTTPError as e:
-                if e.code != HTTPStatus.UNAUTHORIZED:
-                    raise
-            else:
-                return token
-
-    headers = dict(BASE_HEADERS)
-    headers['Authorization'] = 'Bearer %s' % get_token()['access_token']
-
-    spec_resp = open_url(hostname + SPEC_PATH, method=HTTPMethod.GET, headers=headers, validate_certs=False).read()
-    docs_resp = open_url(hostname + DOC_PATH, method=HTTPMethod.GET, headers=headers, validate_certs=False).read()
-    return FdmSwaggerParser().parse_spec(json.loads(to_text(spec_resp)), json.loads(to_text(docs_resp)))
-
-
-class ApiSpecAutocomplete(object):
-    """Allow us to extend API spec with custom actions implemented at Module level.
-
-    Should be extended with _check_and_generate_<operation_type>_specs methods - each of which can extend api spec
-    with specific operation types.
-    The following methods should be called in case of 'api_spec' update required
-        _add_operation_to_model_spec
-        _add_operation_to_operations_spec
-    """
-    UPSERT_DEFAUL_FILTER_COMMENT = " Default filtering for Upsert operation is done by name."
-
-    def __init__(self, api_spec):
-        self._api_spec = api_spec
-        self._operation_checker = OperationChecker
-
-    def _add_operation_to_model_spec(self, model_name, op_name, op_spec):
-        self._api_spec[SpecProp.MODEL_OPERATIONS][model_name][op_name] = op_spec
-
-    def _add_operation_to_operations_spec(self, op_name, op_spec):
-        self._api_spec[SpecProp.OPERATIONS][op_name] = op_spec
-
-    def _generate_upsert_spec(self, operations, model_name, add_operantion, list_operation):
-        op_spec = dict(operations[add_operantion])
-        base_filter_spec = dict(
-            operations[list_operation][OperationField.PARAMETERS][OperationParams.QUERY][QueryParams.FILTER])
-
-        def property_for_filtering_is_present(model_spec, prop_name):
-            """Check that required property is present in the model spec."""
-            return model_spec[PropName.PROPERTIES].get(prop_name)
-
-        if property_for_filtering_is_present(model_spec=self._api_spec[SpecProp.MODELS][model_name],
-                                             prop_name=PropName.NAME):
-            base_filter_spec[PropName.DESCRIPTION] += self.UPSERT_DEFAUL_FILTER_COMMENT
-        else:
-            base_filter_spec[PropName.REQUIRED] = True
-
-        op_spec[OperationField.PARAMETERS][OperationParams.QUERY][QueryParams.FILTER] = base_filter_spec
-
-        return op_spec
-
-    def _check_and_generate_upsert_specs(self, model_name, operations):
-        """Update api spec if upsert action is applicable to model."""
-        # We assume that Upsert and Add actions signature will be same for the moment
-        if not model_name:
-            # Some actions will have no model - we can't do upsert for them.
-            return
-
-        add_operantion = OperationNamePrefix.ADD + model_name
-        list_operation = OperationNamePrefix.GET + model_name + 'List'
-
-        if self._operation_checker.is_upsert_operation_supported(operations):
-            op_name = OperationNamePrefix.UPSERT + model_name
-            op_spec = self._generate_upsert_spec(operations, model_name, add_operantion, list_operation)
-
-            self._add_operation_to_model_spec(
-                model_name=model_name,
-                op_name=op_name,
-                op_spec=op_spec
-            )
-            self._add_operation_to_operations_spec(
-                op_name=op_name,
-                op_spec=op_spec
-            )
-
-    def lookup_and_complete(self):
-        for model_name, operations in self._api_spec[SpecProp.MODEL_OPERATIONS].items():
-            self._check_and_generate_upsert_specs(model_name, operations)
-
-
-def main():
-    def prepare_dist_dir_with_static_docs():
-        if os.path.exists(args.dist):
-            shutil.rmtree(args.dist)
-        shutil.copytree(DEFAULT_STATIC_DIR, args.dist)
-
-    parser = argparse.ArgumentParser(description='Generates Ansible modules from Swagger documentation')
-    parser.add_argument('hostname', type=str, help='Hostname where FTD can be accessed')
-    parser.add_argument('username', type=str, help='FTD username that has access to Swagger docs')
-    parser.add_argument('password', type=str, help='Password for the username')
-    parser.add_argument('--models', type=str, nargs='+', help='A list of models to include in the docs', required=False)
-    parser.add_argument('--dist', type=str, help='An output directory for distribution files', required=False,
-                        default=DEFAULT_DIST_DIR)
-    args = parser.parse_args()
-
-    api_spec = fetch_api_specs_with_docs(args.hostname, args.username, args.password)
-    prepare_dist_dir_with_static_docs()
-
-    spec_autocomplete = ApiSpecAutocomplete(api_spec)
-    spec_autocomplete.lookup_and_complete()
-
-    model_generator = ModelDocGenerator(DEFAULT_TEMPLATE_DIR, api_spec)
-    model_generator.generate_doc_files(args.dist, args.models)
-
-    op_generator = OperationDocGenerator(DEFAULT_TEMPLATE_DIR, api_spec)
-    op_generator.generate_doc_files(args.dist, args.models)
-
-    module_generator = ModuleDocGenerator(DEFAULT_TEMPLATE_DIR, DEFAULT_MODULE_DIR)
-    module_generator.generate_doc_files(args.dist)
-
-    example_generator = ExampleDocGenerator(DEFAULT_TEMPLATE_DIR, DEFAULT_SAMPLES_DIR)
-    example_generator.generate_doc_files(args.dist)
-
-
-if __name__ == '__main__':
-    main()
